Separate test and src
