+++ docker-compose.yml
@@ -15,14 +15,14 @@ services:
       test: ["CMD-SHELL", "pg_isready -U pg -d pg"]
       interval: 1s
       timeout: 1s
-  # api:
-  #   image: api
-  #   build: .
-  #   depends_on:
-  #     db:
-  #       condition: service_healthy
-  #   ports:
-  #     - "8282:8282"
-  #   environment:
-  #     - DATABASE_URL=postgres://pg:pg@db:5432/pg
-  #   command: ["sh", "-c", "./backend"]
+  api:
+    image: api
+    build: .
+    depends_on:
+      db:
+        condition: service_healthy
+    ports:
+      - "8282:8282"
+    environment:
+      - DATABASE_URL=postgres://pg:pg@db:5432/pg
+    command: ["sh", "-c", "./backend"]
