+++ src/main.go
@@ -3,8 +3,8 @@ package main
 import (
 	"fmt"
 
-	"github.com/mingyuanc/GovTech-Technical/routes"
-	"github.com/mingyuanc/GovTech-Technical/utils"
+	"github.com/mingyuanc/GovTech-Technical/src/routes"
+	"github.com/mingyuanc/GovTech-Technical/src/utils"
 )
 
 func main() {
